Use EventDriven.CQRS.Abstractions v2.

+++ reference-architecture/OrderService/Domain/OrderAggregate/CommandHandlers/RemoveOrderHandler.cs
@@ -1,8 +1,8 @@
-using EventDriven.DDD.Abstractions.Commands;
+using EventDriven.CQRS.Abstractions.Commands;
 using OrderService.Domain.OrderAggregate.Commands;
 using OrderService.Repositories;
 
-namespace OrderService.Domain.OrderAggregate.Handlers;
+namespace OrderService.Domain.OrderAggregate.CommandHandlers;
 
 public class RemoveOrderHandler : ICommandHandler<RemoveOrder>
 {
@@ -19,6 +19,15 @@ public class RemoveOrderHandler : ICommandHandler<RemoveOrder>
 
     public async Task<CommandResult> Handle(RemoveOrder command, CancellationToken cancellationToken)
     {
+        // Process command
+        _logger.LogInformation("Handling command: {CommandName}", nameof(RemoveOrder));
+        var entity = await _repository.GetAsync(command.EntityId);
+        if (entity == null) return new CommandResult<Order>(CommandOutcome.NotFound);
+        var domainEvent = entity.Process(command);
+            
+        // Apply events
+        entity.Apply(domainEvent);
+
         // Persist entity
         _logger.LogInformation("Handling command: {CommandName}", nameof(RemoveOrder));
         await _repository.RemoveAsync(command.EntityId);
